Add missing file; rename tests
