+++ src/test-nunit-summary.exe/BriefHtmlOutputTests_NUnit2.cs
@@ -20,11 +20,13 @@
 // ***********************************************************************
 
 using System;
+using System.Text.RegularExpressions;
+using System.Xml;
 using NUnit.Framework;
 
 namespace NUnit.Extras.Tests
 {
-    public class BriefTextOutputTests : ReportCreationTests
+    public class BriefHtmlOutputTests_NUnit2 : ReportCreationTests
     {
         protected override string Input
         {
@@ -33,29 +35,31 @@ namespace NUnit.Extras.Tests
 
         protected override string Output
         {
-            get { return "MockAssemblyBriefSummary-2.6.4.txt"; }
+            get { return "MockAssemblyBriefSummary-2.6.4.html"; }
         }
 
         protected override string Options
         {
-            get { return "-brief"; }
+            get { return "-brief -html"; }
         }
 
         static TestCaseData[] ExpectedText = new TestCaseData[]
         {
             new TestCaseData(@"C:\Program Files\NUnit 2.6.4\bin\tests\mock-assembly.dll"),
-            new TestCaseData("NUnit Version 2.6.4.14350  2016-09-10  11:13:55"),
+            new TestCaseData("NUnit Version: 2.6.4.14350"),
+            new TestCaseData("Date: 2016-09-10"),
+            new TestCaseData("Time: 11:13:55"),
             new TestCaseData("Runtime Environment -"),
-            new TestCaseData("   OS Version: Microsoft Windows NT 6.2.9200.0"),
-            new TestCaseData("  CLR Version: " + Environment.Version),
+            new TestCaseData("OS Version: Microsoft Windows NT 6.2.9200.0"),
+            new TestCaseData("CLR Version: " + Environment.Version),
             new TestCaseData("Tests run: 21, Errors: 1, Failures: 1, Inconclusive: 1, Time: 0.088 seconds"),
-            new TestCaseData("  Not run: 7, Invalid: 3, Ignored: 4, Skipped: 0")
+            new TestCaseData("Not run: 7, Invalid: 3, Ignored: 4, Skipped: 0")
         };
 
         [TestCaseSource("ExpectedText")]
         public void CheckReportContent(string text)
         {
-            Assert.That(Report, Contains.Substring(text));
+            Assert.That(StrippedReport, Contains.Substring(text));
         }
     }
 }
