+++ src/test-nunit-summary.exe/test-nunit-summary.exe.csproj
@@ -49,12 +49,12 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="BriefHtmlOutputTestsV3.cs" />
-    <Compile Include="DefaultHtmlOutputTests.cs" />
-    <Compile Include="BriefHtmlOutputTests.cs" />
+    <Compile Include="BriefHtmlOutputTests_NUnit3.cs" />
+    <Compile Include="DefaultHtmlOutputTests_NUnit2.cs" />
+    <Compile Include="BriefHtmlOutputTests_NUnit2.cs" />
     <Compile Include="ReportCreationTests.cs" />
-    <Compile Include="DefaultTextOutputTests.cs" />
-    <Compile Include="BriefTextOutputTests.cs" />
+    <Compile Include="DefaultTextOutputTests_NUnit2.cs" />
+    <Compile Include="BriefTextOutputTests_NUnit2.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="XmlTransformerOptionsTests.cs" />
